requirements.txt
requests
python-telegram-bot
google-cloud-storage
google-cloud-firestore

flask

flask_login
secret

joblib

numpy
pandas
matplotlib
scikit-learn
plotly

datatime
